fix(migration): preserve TimetableTemplate data on rename
Replace DROP TABLE + CREATE TABLE with ALTER TABLE RENAME TO for the
Template → TimetableTemplate migration. The old Prisma-generated SQL
would have deleted all timetable template rows on production deploy.

Also moves the _utils/time-utils shared module and adds the synchronous
offset-clamp correction to the template editor client.
